Merging Nexson translation and validation

+++ ws-tests/2globalconf.sh
@@ -1,2 +1,7 @@
 #!/bin/sh
-cp global.test.conf test.conf
+if test -f global.conf
+then
+    cp global.conf test.conf
+else
+    cp global.test.conf test.conf
+fi
